Further work with feature bits.
